refactor: lift mapping of status to intent from the healthbar to consumer
We don't want the generic healthbar in ui-toolkit to depend on the Vega
types lib.
