+++ apps/liquidity-provision-dashboard/src/app/components/dashboard/market-list/market-list.tsx
@@ -30,6 +30,7 @@ import { useCallback, useState } from 'react';
 import { Grid } from '../../grid';
 import { HealthDialog } from '../../health-dialog';
 import { Status } from '../../status';
+import { intentForStatus } from '../../../lib/utils';
 import { formatDistanceToNow } from 'date-fns';
 
 export const MarketList = () => {
@@ -273,13 +274,13 @@ export const MarketList = () => {
               data: Market;
             }) => (
               <HealthBar
-                status={value}
                 target={data.target}
                 decimals={
                   data.tradableInstrument.instrument.product.settlementAsset
                     .decimals
                 }
                 levels={data.feeLevels}
+                intent={intentForStatus(value)}
               />
             )}
             sortable={false}
